+++ containers/tefca-viewer/src/app/format-service.tsx
@@ -6,6 +6,8 @@ import {
   ContactPoint,
   Identifier,
 } from "fhir/r4";
+import { ValueSetItem } from "./constants";
+import { QueryStruct } from "./demoQueries";
 
 /**
  * Formats a string.
@@ -257,3 +259,39 @@ export async function GetPhoneQueryFormats(phone: string) {
   });
   return possibleFormats;
 }
+
+/**
+ * Formats a statefully updated list of value set items into a JSON structure
+ * used for executing custom queries.
+ * @param useCase The base use case being queried for.
+ * @param vsItems The list of value set items the user wants included.
+ * @returns A structured specification of a query that can be executed.
+ */
+export const formatValueSetItemsAsQuerySpec = async (
+  useCase: string,
+  vsItems: ValueSetItem[],
+) => {
+  let secondEncounter: boolean = false;
+  if (["cancer", "chlamydia", "gonorrhea", "syphilis"].includes(useCase)) {
+    secondEncounter = true;
+  }
+  const labCodes: string[] = vsItems
+    .filter((vs) => vs.system === "http://loinc.org")
+    .map((vs) => vs.code);
+  const snomedCodes: string[] = vsItems
+    .filter((vs) => vs.system === "http://snomed.info/sct")
+    .map((vs) => vs.code);
+  const rxnormCodes: string[] = vsItems
+    .filter((vs) => vs.system === "http://www.nlm.nih.gov/research/umls/rxnorm")
+    .map((vs) => vs.code);
+
+  const spec: QueryStruct = {
+    labCodes: labCodes,
+    snomedCodes: snomedCodes,
+    rxnormCodes: rxnormCodes,
+    classTypeCodes: [] as string[],
+    hasSecondEncounterQuery: secondEncounter,
+  };
+
+  return spec;
+};
